add methods

+++ src/exam/project/IShippingCareStrategy/IShippingCareStrategy.java
@@ -1,5 +1,5 @@
 package exam.project.IShippingCareStrategy;
 
 public interface IShippingCareStrategy {
-    float CalculateShippingCare(int shippingMultiplier, float weight, float size);
+    double CalculateShippingCare(int shippingMultiplier, double weight, double size);
 }
